requirements.txt
# JSON RPC from bitcoind
python-bitcoinrpc

# ZeroMQ
pyzmq
